pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Kuriko Iwai"]
packages = [{include = "src"}, {include = "tests"}, {include = "Prompts"}]


[tool.poetry.dependencies]
python = "^3.12"
supabase = "^2.7.4"
python-dotenv = "^1.0.1"
uvicorn = "^0.15.0"
together = "^1.2.12"
fastapi = "^0.114.2"
asyncpg = "^0.29.0"
sqlalchemy = "^2.0.34"
pydantic = "^2.9.1"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"



[tool.poetry.scripts]
neg = "tests.test_negotiation:main"
client = "tests.test_endpoint:test_browse_endpoint"
